+++ src/main/java/insane96mcp/iguanatweaksreborn/module/mining/blockhardness/BlockHardness.java
@@ -22,7 +22,7 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 import java.util.ArrayList;
 import java.util.List;
 
-@Label(name = "Block Hardness", description = "Change blocks hardness. Dimension Hardness and Depth Hardness are controlled via json in this feature's folder. Block hardness can be changed with Block Data via data packs")
+@Label(name = "Block Hardness", description = "Change blocks hardness. Dimension Hardness and Depth Hardness are controlled via json in this feature's folder. Block hardness can be changed with Block Definitions via data packs")
 @LoadFeature(module = Modules.Ids.MINING)
 public class BlockHardness extends JsonFeature {
 	public static final TagKey<Block> HARDNESS_BLACKLIST = ISOBlockTagsProvider.create("hardness_blacklist");